Former-commit-id: c7eaf1731810939218038e9d66c69a978a2e199c

+++ src/com/vimukti/accounter/web/client/ui/grids/VendorTransactionGrid.java
@@ -900,7 +900,7 @@ public class VendorTransactionGrid extends
 				try {
 					ClientQuantity quant = new ClientQuantity();
 					if (AccounterValidator.isValidGridQuantity(quantity)) {
-						quant.setValue(Integer.parseInt(qty));
+						quant.setValue(Double.parseDouble(qty));
 						item.setQuantity(quant);
 						update_quantity_inAllRecords(item.getQuantity()
 								.getValue());
